+++ tests/integration/test_installation.py
@@ -1,13 +1,9 @@
 import logging
 import os
 import random
-import shutil
-import subprocess
-import sys
 from io import BytesIO
 from pathlib import Path
 
-import pytest
 from databricks.sdk.service import compute, jobs
 from databricks.sdk.service.iam import PermissionLevel
 from databricks.sdk.service.workspace import ImportFormat
@@ -16,88 +12,8 @@ from databricks.labs.ucx.config import GroupsConfig, MigrationConfig, TaclConfig
 from databricks.labs.ucx.hive_metastore.grants import Grant
 from databricks.labs.ucx.hive_metastore.tables import Table
 from databricks.labs.ucx.install import Installer
-from databricks.labs.ucx.mixins.compute import CommandExecutor
-from databricks.labs.ucx.mounts.list_mounts import Mount
 
 logger = logging.getLogger(__name__)
-
-
-@pytest.fixture
-def fresh_wheel_file(tmp_path) -> Path:
-    this_file = Path(__file__)
-    project_root = this_file.parent.parent.parent.absolute()
-    # TODO: we can dynamically determine this with python -m build .
-    wheel_name = "databricks_labs_ucx"
-
-    build_root = tmp_path / fresh_wheel_file.__name__
-    shutil.copytree(project_root, build_root)
-    try:
-        completed_process = subprocess.run(
-            [sys.executable, "-m", "pip", "wheel", "."],
-            capture_output=True,
-            cwd=build_root,
-            check=True,
-        )
-        if completed_process.returncode != 0:
-            raise RuntimeError(completed_process.stderr)
-
-        found_wheels = list(build_root.glob(f"{wheel_name}-*.whl"))
-        if not found_wheels:
-            msg = f"cannot find {wheel_name}-*.whl"
-            raise RuntimeError(msg)
-        if len(found_wheels) > 1:
-            conflicts = ", ".join(str(whl) for whl in found_wheels)
-            msg = f"more than one wheel match: {conflicts}"
-            raise RuntimeError(msg)
-        wheel_file = found_wheels[0]
-
-        return wheel_file
-    except subprocess.CalledProcessError as e:
-        raise RuntimeError(e.stderr) from None
-
-
-@pytest.fixture
-def wsfs_wheel(ws, fresh_wheel_file, make_random):
-    my_user = ws.current_user.me().user_name
-    workspace_location = f"/Users/{my_user}/wheels/{make_random(10)}"
-    ws.workspace.mkdirs(workspace_location)
-
-    wsfs_wheel = f"{workspace_location}/{fresh_wheel_file.name}"
-    with fresh_wheel_file.open("rb") as f:
-        ws.workspace.upload(wsfs_wheel, f, format=ImportFormat.AUTO)
-
-    yield wsfs_wheel
-
-    ws.workspace.delete(workspace_location, recursive=True)
-
-
-def test_this_wheel_installs(ws, wsfs_wheel):
-    commands = CommandExecutor(ws)
-
-    commands.install_notebook_library(f"/Workspace{wsfs_wheel}")
-    installed_version = commands.run(
-        """
-        from databricks.labs.ucx.__about__ import __version__
-        print(__version__)
-        """
-    )
-
-    assert installed_version is not None
-
-
-def test_sql_backend_works(ws, wsfs_wheel):
-    commands = CommandExecutor(ws)
-
-    commands.install_notebook_library(f"/Workspace{wsfs_wheel}")
-    database_names = commands.run(
-        """
-        from databricks.labs.ucx.framework.crawlers import RuntimeBackend
-        backend = RuntimeBackend()
-        return backend.fetch("SHOW DATABASES")
-        """
-    )
-
-    assert len(database_names) > 0
 
 
 def test_assessment_job_with_no_inventory_database(
@@ -396,29 +312,3 @@ def test_toolkit_notebook(
         logger.info("deleting job")
 
         ws.jobs.delete(created_job.job_id)
-
-
-def test_mount_listing(ws, wsfs_wheel, make_schema, sql_fetch_all):
-    _, inventory_database = make_schema(catalog="hive_metastore").split(".")
-    commands = CommandExecutor(ws)
-    commands.install_notebook_library(f"/Workspace{wsfs_wheel}")
-
-    commands.run(
-        f"""
-        from databricks.labs.ucx.mounts import MountMigrationToolKit
-        from databricks.labs.ucx.config import MigrationConfig, GroupsConfig, TaclConfig
-        cfg = MigrationConfig(
-            inventory_database="{inventory_database}",
-            groups=GroupsConfig(auto=True),
-            tacl=TaclConfig(databases=["default"]))
-        MountMigrationToolKit(cfg).inventorize_mounts()
-        """
-    )
-    mounts = sql_fetch_all(f"SELECT * FROM hive_metastore.{inventory_database}.mounts")
-    results = []
-
-    for mount in mounts:
-        mount_info = Mount(*mount)
-        results.append(mount_info)
-
-    assert len(results) > 0
